+++ app/src/main/java/com/azimolabs/mobile/task/GithubAppComponent.java
@@ -2,7 +2,7 @@ package com.azimolabs.mobile.task;
 
 import com.azimolabs.mobile.task.github.GitHubApiClientModule;
 import com.azimolabs.mobile.task.main.MainActivityComponent;
-import com.azimolabs.mobile.task.main.UserProfileActivityComponent;
+import com.azimolabs.mobile.task.di.UserProfileActivityComponent;
 
 import dagger.Component;
