+++ app/src/main/res/layout/activity_user_profile.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_gravity="center"
-    tools:context=".main.UserProfileActivity">
+    tools:context=".view.UserProfileActivity">
 
     <ProgressBar
         android:id="@+id/pbLoadingRepos"
@@ -30,7 +30,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:background="@color/colorPrimary"
-            android:elevation="6dp"
+            android:elevation="4dp"
             android:paddingBottom="24dp">
 
             <TextView
